# agent-backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "UserController_addUser",
        "summary": "Create a new user",
        "description": "This endpoint creates a new user in the system with the provided user details.",
        "tags": ["User"],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Response returned after successful user creation.",
                  "properties": {
                    "status": {
                      "type": "string",
                      "description": "Status of the operation.",
                      "example": "SUCCESS"
                    },
                    "data": {
                      "type": "object",
                      "description": "Created user data.",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "description": "User ID.",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        },
                        "name": {
                          "type": "string",
                          "description": "Full name of the user.",
                          "example": "Jane Doe"
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "description": "Email address of the user.",
                          "example": "user@example.com"
                        },
                        "isEmailVerified": {
                          "type": "boolean",
                          "description": "Whether the user's email is verified.",
                          "example": false
                        },
                        "account": {
                          "type": "object",
                          "description": "Associated account info.",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "uuid",
                              "example": "123e4567-e89b-12d3-a456-426614174000"
                            },
                            "name": {
                              "type": "string",
                              "example": "Example Account"
                            }
                          }
                        },
                        "role": {
                          "type": "object",
                          "description": "User role information.",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "uuid",
                              "example": "123e4567-e89b-12d3-a456-426614174000"
                            },
                            "code": {
                              "type": "string",
                              "example": "admin"
                            },
                            "value": {
                              "type": "string",
                              "example": "Admin"
                            }
                          }
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "description": "User creation timestamp.",
                          "example": "2025-07-25T07:31:42.431Z"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time",
                          "description": "Last updated timestamp.",
                          "example": "2025-07-25T07:31:42.431Z"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Timestamp of the response.",
                      "example": "2025-07-25T07:31:42.470Z"
                    }
                  }
                },
                "example": {
                  "status": "SUCCESS",
                  "data": {
                    "id": "cf55b867-8098-4925-8d23-8052b44bb455",
                    "name": "Jane Doe",
                    "email": "user@example.com",
                    "isEmailVerified": false,
                    "account": {
                      "id": "d3a7a16c-2b99-4be6-9b91-8f75a21e9a5d",
                      "name": "Example Account"
                    },
                    "role": {
                      "id": "8c4bc16b-49b0-4ffa-935a-6238ba025db0",
                      "code": "admin",
                      "value": "Admin"
                    },
                    "createdAt": "2025-07-25T07:31:42.431Z",
                    "updatedAt": "2025-07-25T07:31:42.431Z"
                  },
                  "timestamp": "2025-07-25T07:31:42.470Z"
                }
              }
            }
          },
          "400": {
            "description": "User creation failed due to invalid input or server-side error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Failure response format.",
                  "properties": {
                    "status": {
                      "type": "string",
                      "description": "Failure status indicator.",
                      "example": "FAILURE"
                    },
                    "message": {
                      "type": "string",
                      "description": "Error message for the failure.",
                      "example": ""
                    },
                    "errorCode": {
                      "type": "string",
                      "description": "Error code for debugging.",
                      "example": "FAILED_USER_UPSERTION"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Time of the error.",
                      "example": "2025-07-25T07:48:22.152Z"
                    },
                    "path": {
                      "type": "string",
                      "description": "Endpoint where the error occurred.",
                      "example": "/agent-backend/users"
                    }
                  }
                },
                "example": {
                  "status": "FAILURE",
                  "message": "",
                  "errorCode": "FAILED_USER_UPSERTION",
                  "timestamp": "2025-07-25T07:48:22.152Z",
                  "path": "/agent-backend/users"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "UserController_updateUser",
        "summary": "Update user details",
        "description": "Updates a user's details based on the request body.",
        "tags": ["User"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "SUCCESS"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "isEmailVerified": {
                          "type": "boolean"
                        },
                        "account": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "name": {
                              "type": "string"
                            }
                          }
                        },
                        "role": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "code": {
                              "type": "string"
                            },
                            "value": {
                              "type": "string"
                            }
                          }
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "example": {
                    "status": "SUCCESS",
                    "data": {
                      "id": "cf55b867-8098-4925-8d23-8052b44bb455",
                      "name": "Pratik Ghodke",
                      "email": "pratik.ghodke@sarvaha.com",
                      "isEmailVerified": false,
                      "account": {
                        "id": "d3a7a16c-2b99-4be6-9b91-8f75a21e9a5d",
                        "name": "Sarvaha Systems"
                      },
                      "role": {
                        "id": "8c4bc16b-49b0-4ffa-935a-6238ba025db0",
                        "code": "admin",
                        "value": "Admin"
                      },
                      "createdAt": "2025-07-25T07:31:42.431Z",
                      "updatedAt": "2025-07-25T08:00:39.609Z"
                    },
                    "timestamp": "2025-07-25T08:00:39.642Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or incorrect information provided.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "FAILURE"
                    },
                    "message": {
                      "type": "string",
                      "example": "INCORRECT_INFORMATION_PROVIDED"
                    },
                    "errorCode": {
                      "type": "string",
                      "example": "INVALID_INPUT"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "path": {
                      "type": "string",
                      "example": "/agent-backend/users"
                    }
                  }
                },
                "example": {
                  "status": "FAILURE",
                  "message": "INCORRECT_INFORMATION_PROVIDED",
                  "errorCode": "INVALID_INPUT",
                  "timestamp": "2025-07-25T08:00:59.680Z",
                  "path": "/agent-backend/users"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "UserController_getUserByAccount",
        "summary": "Get users by account",
        "description": "Fetch a list of users associated with an account.",
        "tags": ["User"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Users retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "SUCCESS"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "numberOfRows": {
                          "type": "integer",
                          "example": 1
                        },
                        "rows": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid",
                                "example": "123e4567-e89b-12d3-a456-426614174000"
                              },
                              "name": {
                                "type": "string",
                                "example": "Jane Doe"
                              },
                              "email": {
                                "type": "string",
                                "format": "email",
                                "example": "user@example.com"
                              },
                              "password": {
                                "type": "string",
                                "description": "Encrypted password (internal use only)",
                                "example": "encrypted_password_here"
                              },
                              "isEmailVerified": {
                                "type": "boolean",
                                "example": false
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-07-25T07:31:42.431Z"
                              },
                              "updatedAt": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-07-25T08:00:39.609Z"
                              },
                              "role": {
                                "type": "object",
                                "properties": {
                                  "code": {
                                    "type": "string",
                                    "example": "admin"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-07-25T08:36:31.455Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required input.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "FAILURE"
                    },
                    "message": {
                      "type": "string",
                      "example": "Either id or accountId is required to fetch the user."
                    },
                    "errorCode": {
                      "type": "string",
                      "example": "MISSING_INPUT"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-07-25T08:38:51.717Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/agent-backend/users/"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/documents/{id}": {
      "put": {
        "operationId": "DocumentController_updateDocument",
        "summary": "Update document metadata",
        "description": "Updates metadata for a given document by its ID.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document updated successfully.",
            "content": {
              "application/json": {
                "example": {
                  "status": "SUCCESS",
                  "data": {
                    "fileName": "updated-file-name.ext",
                    "id": "{documentId}",
                    "userId": "{userId}",
                    "accountId": "{accountId}",
                    "folder": "{folderPath}",
                    "type": "file",
                    "status": "{status}",
                    "createdAt": "2025-07-25T11:45:14.740Z",
                    "updatedAt": "2025-07-25T12:07:54.792Z"
                  },
                  "timestamp": "2025-07-25T12:07:54.802Z"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden access. The user does not have permission to update this document.",
            "content": {
              "application/json": {
                "example": {
                  "status": "FAILURE",
                  "message": "FORBIDDEN_ACCESS",
                  "errorCode": "FORBIDDEN_ACCESS",
                  "timestamp": "2025-07-25T12:09:06.842Z",
                  "path": "/agent-backend/documents/{documentId}"
                }
              }
            }
          }
        },
        "tags": ["Document"]
      }
    },
...
  },
  "info": {
    "title": "Agent Backend Service",
    "description": "✨The Agent Backend Service is a scalable, modular backend system designed to power intelligent agent-based applications. It provides robust APIs for authentication, user management, document processing, conversation tracking, training resources, and more. Built with flexibility in mind, it integrates seamlessly with external services while ensuring security, extensibility, and high availability.\n\n# 🔐 Authentication\n\nThis API uses a secure cookie named `token` to authenticate requests.\n\n- You must log in to receive this cookie.\n- When using **Try It Out** in the docs, **you must manually set the `token` cookie** in your browser.\n- You can find it in dev tools → Application → Cookies → {current url} -> `token`.\n\n**Note**: Scalar does not automatically send cookies during \"Try it out\".\n",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "name": "prod",
      "url": "https://zeno.sarvaha.ai/agent-backend"
    },
    {
      "name": "dev",
      "url": "http://localhost:7082/agent-backend"
    }
  ],
  "components": {
    "schemas": {
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "accountId": {
            "type": "string",
            "format": "uuid",
            "example": "987f6543-e21a-45bc-94d1-112233445566"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "nextLoginDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-08-01T10:00:00Z"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 5,
            "example": "yourStrongPassword123"
          },
          "isDeleted": {
            "type": "boolean",
            "example": false
          },
          "isEmailVerified": {
            "type": "boolean",
            "example": true
          },
          "roleId": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        }
      },
...
      "UpdateDocumentDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "example": "document.pdf"
          },
          "status": {
            "type": "string",
            "enum": ["completed", "failed", "new", "uploaded", "in progress"],
            "example": "completed"
          },
          "isDeleted": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
    }
  }
}